+++ tests/integrations/datasets/dataset_test.py
@@ -1,13 +1,13 @@
 from tango.common.testing import TangoTestCase
-from tango.integrations.datasets import HuggingFaceDataset
+from tango.integrations.datasets import LoadDataset
 from tango.step import Step
 
 
-class TestHuggingFaceDataset(TangoTestCase):
+class TestLoadDataset(TangoTestCase):
     def test_from_params(self):
-        step: HuggingFaceDataset = Step.from_params(  # type: ignore[assignment]
+        step: LoadDataset = Step.from_params(  # type: ignore[assignment]
             {
-                "type": "hf_dataset",
+                "type": "datasets::load",
                 "path": "lhoestq/test",
                 "cache_dir": str(self.TEST_DIR / "cache"),
             }